requirements.txt
dgl
matplotlib
numpy
pytorch
gensim
mecab-python3
nltk
